Revert "Development Updates (#6)"
This reverts commit 6372556ccedd66f563b97bfe19a26733dc1495f5.
