+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 anyhow = "1.0.66"
 shuttle-serenity = "0.25.0"
 shuttle-runtime = "0.25.0"
-serenity = { version = "0.11.5", default-features = false, features = ["client", "gateway", "rustls_backend", "model", "cache"] }
+serenity = { version = "0.11.5", default-features = false, features = ["client", "gateway", "rustls_backend", "model"] }
 shuttle-secrets = "0.25.0"
 tokio = "1.26.0"
 tracing = "0.1.37"
@@ -15,4 +15,4 @@ reqwest = "0.11.20"
 serde = "1.0.188"
 shuttle-turso = "0.25.0"
 libsql-client = "0.30.1"
-#poise = "0.5.5"
+poise = "0.5.5"
